Extract SimpleHTTPServer and with_http to TestOpenURIUtils module for other tests
